convert to bits at the encoding process
